Added proper jump animation, cleaned up a lot of code to function much better with new input, changed a lot of static values to better fit the real game, added mouse input for attacking as another option.
NOTE: There is an issue with at least my keyboard, likely others, with the n-key input of using left and up arrow keys whilst trying to shoot using space. This is remedied with the addition of the mouse and also using WASD keys works fine here too, but something to bring up that I ran into this issue pretty early on, but there's not a whole lot to do without making hardcoded exceptions everywhere.